+++ ddcUtils/__init__.py
@@ -1,15 +1,17 @@
 import logging
 from importlib.metadata import version
 from typing import Literal, NamedTuple
+from .conf_file_utils import ConfFileUtils
 from .crypto import Cryptography
 from .exceptions import get_exception
 from .file_utils import FileUtils
-from .log import TimedRotatingLog, SizeRotatingLog
+from .log import SizeRotatingLog, TimedRotatingLog
 from .misc_utils import MiscUtils, Object
 from .os_utils import OsUtils
 
 
 __all__ = (
+    "ConfFileUtils",
     "Cryptography",
     "get_exception",
     "FileUtils",
