+++ ddcUtils/log.py
@@ -3,9 +3,9 @@ import gzip
 import logging.handlers
 import os
 import sys
+from logging.handlers import RotatingFileHandler, TimedRotatingFileHandler
 from .exceptions import get_exception
 from .file_utils import FileUtils
-from logging.handlers import RotatingFileHandler, TimedRotatingFileHandler
 
 
 class TimedRotatingLog:
@@ -24,7 +24,7 @@ class TimedRotatingLog:
     def __init__(
         self,
         level: str = "info",
-        directory: str = "./logs",
+        directory: str = "logs",
         filename: str = "app.log",
         encoding: str = "UTF-8",
         days_to_keep: int = 7,
@@ -46,7 +46,9 @@ class TimedRotatingLog:
                                              when=self.when,
                                              utc=self.utc,
                                              backupCount=self.days_to_keep)
-        return _set_log_format(file_hdlr, self.level, self.directory, self.days_to_keep)
+        file_hdlr.suffix = "%Y%m%d"
+        file_hdlr.rotator = GZipRotatorTimed(self.directory, self.days_to_keep)
+        return _set_log_format(file_hdlr, self.level)
 
 
 class SizeRotatingLog:
@@ -57,7 +59,7 @@ class SizeRotatingLog:
     def __init__(
         self,
         level: str = "info",
-        directory: str = "./logs",
+        directory: str = "logs",
         filename: str = "app.log",
         encoding: str = "UTF-8",
         days_to_keep: int = 7,
@@ -79,67 +81,57 @@ class SizeRotatingLog:
                                         encoding=self.encoding,
                                         delay=False,
                                         errors=None)
-        return _set_log_format(file_hdlr, self.level, self.directory, self.days_to_keep)
+        file_hdlr.rotator = GZipRotatorSize(self.directory, self.days_to_keep)
+        return _set_log_format(file_hdlr, self.level)
 
 
-class GZipTimeRotator:
-    def __init__(self, dir_logs, days_to_keep):
+class GZipRotatorSize:
+    def __init__(self, dir_logs: str, days_to_keep: int):
         self.dir = dir_logs
         self.days_to_keep = days_to_keep
 
-    def __call__(self, source, dest):
+    def __call__(self, source: str, dest: str) -> None:
         RemoveOldLogs(self.dir, self.days_to_keep)
         if os.path.isfile(source) and os.stat(source).st_size > 0:
-            try:
-                sfname, sext = os.path.splitext(source)
-                _, dext = os.path.splitext(dest)
-                renamed_dst = f"{sfname}_{dext.replace('.', '')}{sext}.gz"
-                with open(source, "rb") as fin:
-                    with gzip.open(renamed_dst, "wb") as fout:
-                        fout.writelines(fin)
-                os.remove(source)
-            except Exception as e:
-                _write_stderr(f"[Unable to zip log file]:{get_exception(e)}: {source}")
+            new_file_number = 1
+            old_gz_files_list = FileUtils.list_files(self.dir, ends_with=".gz")
+            if old_gz_files_list:
+                try:
+                    oldest_file_number = int(old_gz_files_list[-1].name.split(".")[0].split("_")[1])
+                    new_file_number = oldest_file_number + 1
+                except ValueError as e:
+                    _write_stderr(f"[Unable to get old zip log file number]:{get_exception(e)}: {old_gz_files_list[-1].name}")
+            _gzip_file(source, new_file_number)
 
 
-class GZipSizeRotator:
-    def __init__(self, dir_logs, days_to_keep):
+class GZipRotatorTimed:
+    def __init__(self, dir_logs: str, days_to_keep: int):
         self.dir = dir_logs
         self.days_to_keep = days_to_keep
 
-    def __call__(self, source, dest):
+    def __call__(self, source: str, dest: str) -> None:
         RemoveOldLogs(self.dir, self.days_to_keep)
-        if os.path.isfile(source) and os.stat(source).st_size > 0:
-            try:
-                sfname, sext = os.path.splitext(source)
-                _, dext = os.path.splitext(dest)
-                renamed_dst = f"{sfname}_{dext.replace('.', '')}{sext}.gz"
-                with open(source, "rb") as fin:
-                    with gzip.open(renamed_dst, "wb") as fout:
-                        fout.writelines(fin)
-                os.remove(source)
-            except Exception as e:
-                _write_stderr(f"[Unable to zip log file]:{get_exception(e)}: {source}")
+        output_dated_name = os.path.splitext(dest)[1].replace(".", "")
+        _gzip_file(source, output_dated_name)
 
 
 class RemoveOldLogs:
-    def __init__(self, logs_dir, days_to_keep):
-        files_list = [f for f in os.listdir(logs_dir)
-                      if os.path.isfile(f"{logs_dir}/{f}") and os.path.splitext(f)[1] == ".gz"]
+    def __init__(self, logs_dir: str, days_to_keep: int) -> None:
+        files_list = FileUtils.list_files(logs_dir, ends_with=".gz")
         for file in files_list:
-            file_path = os.path.join(logs_dir, file)
-            if FileUtils.is_file_older_than_x_days(file_path, days_to_keep):
+            file_path = str(os.path.join(logs_dir, file))
+            if FileUtils.is_older_than_x_days(file_path, days_to_keep):
                 try:
-                    os.remove(file_path)
+                    FileUtils.remove(file_path)
                 except Exception as e:
                     _write_stderr(f"[Unable to remove old logs]:{get_exception(e)}: {file_path}")
 
 
-def _write_stderr(msg):
+def _write_stderr(msg: str) -> None:
     sys.stdout.write(f"[ERROR]:{msg}\n")
 
 
-def _get_level(level: str):
+def _get_level(level: str) -> logging:
     if not isinstance(level, str):
         _write_stderr("[Unable to get log level]. Default level to: 'info'")
         return logging.INFO
@@ -156,14 +148,14 @@ def _get_level(level: str):
             return logging.INFO
 
 
-def _get_log_path(directory, filename):
+def _get_log_path(directory: str, filename: str) -> str:
     try:
         os.makedirs(directory, exist_ok=True) if not os.path.isdir(directory) else None
     except Exception as e:
         _write_stderr(f"[Unable to create logs directory]:{get_exception(e)}: {directory}")
         raise e
 
-    log_file_path = os.path.join(directory, filename)
+    log_file_path = str(os.path.join(directory, filename))
 
     try:
         open(log_file_path, "a+").close()
@@ -171,10 +163,10 @@ def _get_log_path(directory, filename):
         _write_stderr(f"[Unable to open log file for writing]:{get_exception(e)}: {log_file_path}")
         raise e
 
-    return str(log_file_path)
+    return log_file_path
 
 
-def _set_log_format(file_hdlr, level, directory, days_to_keep):
+def _set_log_format(file_hdlr, level: logging) -> logging.Logger:
     _debug_formatt = ""
     if level == logging.DEBUG:
         _debug_formatt = f"[PID:{os.getpid()}]:[%(filename)s:%(funcName)s:%(lineno)d]:"
@@ -186,12 +178,6 @@ def _set_log_format(file_hdlr, level, directory, days_to_keep):
     logger.setLevel(level)
 
     file_hdlr.setFormatter(formatter)
-    if isinstance(file_hdlr, logging.handlers.TimedRotatingFileHandler):
-        file_hdlr.suffix = "%Y%m%d"
-        file_hdlr.rotator = GZipTimeRotator(directory, days_to_keep)
-    else:
-        file_hdlr.rotator = GZipSizeRotator(directory, days_to_keep)
-
     file_hdlr.setLevel(level)
     logger.addHandler(file_hdlr)
 
@@ -201,3 +187,16 @@ def _set_log_format(file_hdlr, level, directory, days_to_keep):
     logger.addHandler(stream_hdlr)
 
     return logger
+
+
+def _gzip_file(source, output_partial_name):
+    if os.path.isfile(source) and os.stat(source).st_size > 0:
+        try:
+            sfname, sext = os.path.splitext(source)
+            renamed_dst = f"{sfname}_{output_partial_name}{sext}.gz"
+            with open(source, "rb") as fin:
+                with gzip.open(renamed_dst, "wb") as fout:
+                    fout.writelines(fin)
+            FileUtils.remove(source)
+        except Exception as e:
+            _write_stderr(f"[Unable to zip log file]:{get_exception(e)}: {source}")
